ocr-worker-paddle/pyproject.toml
[project]
name = "ocr-worker-paddle"
version = "0.1.0"
description = "PaddleOCR GPU worker for the archiver pipeline"
requires-python = ">=3.10"
dependencies = [
    "worker-common",
    "paddleocr>=3.4,<4",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
